marqeta-business-transition-module: abstracted transition logic to separate module
